requirements.txt
# requirements.txt

# Core dependencies
langchain>=0.1.0
pyyaml>=6.0
chromadb>=0.4.22

# Embeddings
sentence-transformers>=2.2.2
InstructorEmbedding>=1.0.1
# For BGE-M3 specifically
torch>=2.0.0
transformers>=4.34.0

# Document loaders
PyPDF2>=3.0.0
unstructured>=0.11.0
markdown>=3.5.1

# For LLMs
accelerate>=0.20.0
einops>=0.6.1
bitsandbytes>=0.41.0  # For efficient quantization
datasets>=2.12.0
huggingface-hub>=0.19.0
safetensors>=0.3.1

# For evaluation
rouge-score>=0.1.2
nltk>=3.8.1
scikit-learn>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0

# For CLI
argparse>=1.4.0

# If you encounter TensorFlow issues with transformers
tf-keras>=2.15.0
colorama>=0.4.6
streamlit>=1.25.0